fix: resolve compiler warnings for clean -Wall -Wextra -Wshadow build
Systematically fix compiler warnings across the FAISS codebase to enable
clean builds with -Wall -Wextra -Wshadow. Changes include:

- Rename shadowing variables (constructor params, loop vars, locals)
- Fix signed/unsigned comparison mismatches
- Add explicit casts where needed (e.g. size_t for OpenMP on MSVC)
- Mark intentionally unused parameters with /*name*/ comments
- Add cmake/StrictCompilerWarnings.cmake for opt-in strict builds

+++ faiss/impl/index_read.cpp
@@ -262,8 +262,9 @@ std::unique_ptr<VectorTransform> read_VectorTransform_up(IOReader* f) {
         READ1(lt->have_bias);
         READVECTOR(lt->A);
         READVECTOR(lt->b);
-        FAISS_THROW_IF_NOT(lt->A.size() >= lt->d_in * lt->d_out);
-        FAISS_THROW_IF_NOT(!lt->have_bias || lt->b.size() >= lt->d_out);
+        FAISS_THROW_IF_NOT(
+                lt->A.size() >= size_t(lt->d_in) * size_t(lt->d_out));
+        FAISS_THROW_IF_NOT(!lt->have_bias || lt->b.size() >= size_t(lt->d_out));
         lt->set_is_orthonormal();
         vt = std::move(lt);
     } else if (h == fourcc("RmDT")) {
@@ -1170,7 +1171,7 @@ std::unique_ptr<Index> read_index_up(IOReader* f, int io_flags) {
                     "invalid IVFFlatDedup instances table size: %zd "
                     "(must be even)",
                     tab.size());
-            for (long i = 0; i < tab.size(); i += 2) {
+            for (size_t i = 0; i < tab.size(); i += 2) {
                 std::pair<idx_t, idx_t> pair(tab[i], tab[i + 1]);
                 ivfl->instances.insert(pair);
             }
@@ -1223,7 +1224,7 @@ std::unique_ptr<Index> read_index_up(IOReader* f, int io_flags) {
         read_ScalarQuantizer(&ivsc->sq, f);
         READ1(ivsc->code_size);
         ArrayInvertedLists* ail = set_array_invlist(ivsc.get(), ids);
-        for (int i = 0; i < ivsc->nlist; i++)
+        for (size_t i = 0; i < ivsc->nlist; i++)
             READVECTOR(ail->codes[i]);
         idx = std::move(ivsc);
     } else if (h == fourcc("IwSQ") || h == fourcc("IwSq")) {
@@ -1730,7 +1731,8 @@ static void read_InvertedLists(IndexBinaryIVF& ivf, IOReader* f, int io_flags) {
     InvertedLists* ils = read_InvertedLists(f, io_flags);
     FAISS_THROW_IF_NOT(
             !ils ||
-            (ils->nlist == ivf.nlist && ils->code_size == ivf.code_size));
+            (ils->nlist == ivf.nlist &&
+             ils->code_size == size_t(ivf.code_size)));
     ivf.invlists = ils;
     ivf.own_invlists = true;
 }
@@ -1828,7 +1830,8 @@ std::unique_ptr<IndexBinary> read_index_binary_up(IOReader* f, int io_flags) {
         auto idxf = std::make_unique<IndexBinaryFlat>();
         read_index_binary_header(*idxf, f);
         read_vector(idxf->xb, f);
-        FAISS_THROW_IF_NOT(idxf->xb.size() == idxf->ntotal * idxf->code_size);
+        FAISS_THROW_IF_NOT(
+                idxf->xb.size() == size_t(idxf->ntotal) * idxf->code_size);
         idx = std::move(idxf);
     } else if (h == fourcc("IBwF")) {
         auto ivf = std::make_unique<IndexBinaryIVF>();
